requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-graphiql-debug-toolbar==0.2.0
django-mptt==0.13.4
django-pglocks==1.0.4
django-prometheus==2.1.0
django-redis==5.0.0
django-rq==2.4.1
django-tables2==2.4.1
django-taggit==1.5.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg[validation]==1.20.0
graphene_django==2.15.0
gunicorn==20.1.0
Jinja2==3.0.2
Markdown==3.3.4
markdown-include==0.6.0
mkdocs-material==7.3.2
netaddr==0.8.0
Pillow==8.3.2
psycopg2-binary==2.9.1
PyYAML==5.4.1
svgwrite==1.4.1
tablib==3.0.0
schema==0.7.4
simplejson==3.17.0
# Workaround for #7401
jsonschema==3.2.0
